remove shadowJar as runtime dependencies are used transitively
We previously assumed, that we had to account for our dependencies on
our own and therefore needed to shade all dependencies into our library
jar before deploying it. It turns out that the implementation
dependencies are retained as transitive dependencies in the
runtimeClasspath of our consumers. Therefore, our dependencies are
included in their Fat-Jars already, and we don't need to provide them
with shaded dependencies and can remove shadowJar from the project.